Used code styler. renamed virtual machien map to just virtual machine
